Move nlu models into subfolder in nlu directory
The superclass intent_classifier.py stays in the nlu
directory. The implementations are supposed to go into the
now created models directory. This will create a clear
separation between implementation and the interface.
